+++ src/app/core/validators/nameNotTaken.validator.ts
@@ -1,7 +1,7 @@
 import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
-import { TemplateService } from '../services/template.service';
+import { IssueTemplateService } from '../services/issue-template.service';
 
-export function nameNotTaken(templateService: TemplateService): ValidatorFn {
+export function nameNotTaken(templateService: IssueTemplateService): ValidatorFn {
   return (name: AbstractControl): ValidationErrors | null => {
     if (templateService.isNameTaken(name.value)) {
       return { taken: true };
