+++ src/app/phase-bug-reporting/new-issue-template/new-issue-template.component.css
