+++ src/app/phase-bug-reporting/phase-bug-reporting.module.ts
@@ -8,7 +8,7 @@ import { SharedModule } from '../shared/shared.module';
 import { ViewIssueModule } from '../shared/view-issue/view-issue.module';
 import { IssueComponent } from './issue/issue.component';
 import { NewIssueComponent } from './new-issue/new-issue.component';
-import { NewTemplateComponent } from './new-template/new-template.component';
+import { NewTemplateComponent } from './new-issue-template/new-issue-template.component';
 import { PhaseBugReportingRoutingModule } from './phase-bug-reporting-routing.module';
 import { PhaseBugReportingComponent } from './phase-bug-reporting.component';
 import { IssuesPostedComponent } from './issues-posted/issues-posted.component';
